app/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# LangChain & LangGraph
langgraph>=0.0.15
langchain>=0.1.0
langgraph-checkpoint-postgres>=0.1.0
langchain-community>=0.0.10
langchain-mcp-adapters>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-deepseek>=0.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0

# LLM Providers
dashscope>=1.19.0

# MCP
mcp>=1.0.0